scripts/requirements.txt
fastapi==0.128.0
numpy==2.0.2
pillow==11.3.0
python-multipart==0.0.20
uvicorn==0.39.0